Update README.md
When duplicating a menu item, an option to automatically copy all assigned modules to the newly created menu item can be added. A checkbox, such as "Duplicate with all assigned modules," could be included when using the "Save as Copy" button. This ensures seamless duplication while maintaining consistent module assignments.