Update README.md
Duplicate Menu Items With Its Module Assignment Option. When duplicating a menu item, an option to automatically copy all assigned modules to the newly created menu item.
When duplicating a menu item, an option to automatically copy all assigned modules to the newly created menu item can be added. A checkbox, such as "Duplicate with all assigned modules," could be included when using the "Save as Copy" button. This ensures seamless duplication while maintaining consistent module assignments.
Actually there was a plugin for joomla 3 that I used to use in the past. Then i ve created one for j4 ( https://extensions.joomla.org/extension/jt-copy-module-assignments/) it works well but i thought that would be great if that feature was added core.